dashboard : separate class (header & popular)
