+++ lib/screens/dashboard_screens/shop_screens/components/detail_product/detail_product_dialog.dart
@@ -4,9 +4,9 @@ import 'package:modal_bottom_sheet/modal_bottom_sheet.dart';
 import 'package:xetia_shop/components/my_button.dart';
 import 'package:xetia_shop/constants.dart';
 import 'package:xetia_shop/core/model/product/product_response.dart';
-import 'package:xetia_shop/screens/dashboard_screens/shop_screens/products_components/detail_product/four_rectangle_button.dart';
-import 'package:xetia_shop/screens/dashboard_screens/shop_screens/products_components/detail_product/title_detail.dart';
-import 'package:xetia_shop/screens/dashboard_screens/shop_screens/products_components/rating_bar.dart';
+import 'package:xetia_shop/screens/dashboard_screens/shop_screens/components/detail_product/four_rectangle_button.dart';
+import 'package:xetia_shop/screens/dashboard_screens/shop_screens/components/detail_product/title_detail.dart';
+import 'package:xetia_shop/screens/dashboard_screens/shop_screens/components/rating_bar.dart';
 
 import 'add_or_remove_product.dart';
