+++ lib/screens/dashboard_screens/shop_screens/components/rating_bar.dart
